refactor(auth): Create separated files for each request
