+++ src/views/_authentication_views.py
@@ -1,16 +1,16 @@
 import flask
-from src.controllers import _authentication_controllers
 from src.middlewares import auth_middlewares
+from src.controllers.authentication import AUTHENTICATION_HANDLERS
 
 views = flask.Blueprint("authentication", __name__)
 
 
 @views.route("/auth/login", methods=["POST"])
 def auth_login():
-    return _authentication_controllers.login_handler()
+    return AUTHENTICATION_HANDLERS["LOGIN"]()
 
 
 @views.route("/auth/refresh", methods=["POST"])
 @auth_middlewares.token_required
 def auth_refresh(token):
-    return _authentication_controllers.challenge_handler(token)
+    return AUTHENTICATION_HANDLERS["REFRESH"](token)
